+++ src/components/ownerNoticeDetail/ApplicationList/ApplicationList.module.scss
