+++ src/components/ownerNoticeDetail/NoticeInformation/NoticeInformation.module.scss
